Add files via upload
This is a simple recommendation model. It is a very generic model primarily based on popularity, voting rate, voting numbers.
Initially, the list is shortlisted based on the quartile values P90, selecting the highest 10% vote_counts.  
Further, the model calculates the weighted score based on voting rate and numbers and sorts the list to get top movies as recommendations to the application.