Add files via upload
Content-Based Recommendation Model:
In this coding scheme, will learn how to build a recommendation model that suggests movies that are similar to a particular movie based upon some content provided.
For the dataset provided, there is content in the 'overview' feature, which is explaining about the movie in textual formatted data in about 20-to-30 words.
Here, the main job is to process the overview content of the movies using 'word vectors' method. Based upon the meaning created by word vectors find similar ones and can be used as a basis for recommendations for a given movie.
In simple words, we calculate the similarity matrix too & which will help us to know the very similar movies for a given input movie title.